refactor(extensions): scope-aware command flows + consolidate shared helpers
Make project-scope extensions correct end-to-end and remove the
bookkeeping that had been duplicated across hk-desktop and hk-web.

* AgentAdapter: add `mcp_config_path_for(scope)` /
  `hook_config_path_for(scope)` trait helpers; collapse 14 duplicated
  match blocks across `manager::toggle_*`, `delete_extension`,
  `get_extension_content`, and `install_to_agent`. Cursor declares
  `.cursor/hooks.json` as its project hook config (per the docs);
  Copilot keeps its dir-scanning behavior via a tiny `read_hooks_from`
  override that delegates to `read_hooks()`.

* hk_core::service: extract `delete_extension`, `get_extension_content`,
  and `install_to_agent` here so the desktop/web wrappers shrink to a
  few lines each. The service fns are scope-aware: project entries read
  from `<project>/<project_*_relpath>`, global from the adapter's
  user-scope path. New `ExtensionContent` type lives alongside.

* hk_core::scanner: consolidate `find_skill_by_id` (was duplicated in
  desktop helpers and web extensions) and make it + `skill_locations`
  project-aware. Project MCP/hook scanner entries now populate
  `source_path` so the detail panel's Paths card can locate them.

* hk_core::store: `list_project_tuples()` helper kills 14 sites of
  `store.list_projects().unwrap_or_default()...map(...).collect()`.

* `check_updates` / `update_extension` skip Project-scope skills so the
  marketplace update flow won't auto-link a user's project skill to a
  marketplace skill that just happens to share a name (and then
  overwrite their copy on the next update).

* hk-desktop/helpers.rs: delete dead `audit_extension_by_name` —
  superseded by `hk_core::service::post_install_sync`, kept alive only
  by `#[allow(dead_code)]`.

343 unit + integration tests pass. Independent code review confirmed
behavioral parity with the pre-refactor wrappers (no I/O while holding
the store mutex; per-kind branches match line-for-line).
