+++ crates/hk-core/src/service.rs
@@ -2,10 +2,12 @@ use crate::{
     HkError,
     adapter::AgentAdapter,
     auditor::{AuditInput, Auditor},
+    deployer,
     models::*,
     scanner,
     store::Store,
 };
+use parking_lot::Mutex;
 
 /// Common post-install flow: scan affected agents, sync to store, set install meta,
 /// update pack, audit the installed extension(s), and persist audit results.
@@ -249,6 +251,602 @@ fn find_skill_content(
         }
     }
     (String::new(), None)
+}
+
+// --- Extension command flows shared by hk-desktop and hk-web -------------
+
+/// Rich detail returned by `get_extension_content`. Surfaces the on-disk
+/// representation (file/dir path + readable text) so the UI's detail panel
+/// can show it. `symlink_target` is only set for skills whose entry or
+/// containing dir is a symlink — useful for development setups where the
+/// user keeps the canonical copy elsewhere.
+#[derive(Debug, Clone, serde::Serialize)]
+pub struct ExtensionContent {
+    pub content: String,
+    pub path: Option<String>,
+    pub symlink_target: Option<String>,
+}
+
+/// Remove an extension from disk/config (per-kind) and then from the DB.
+///
+/// Disk and DB are mutated in two separately-locked phases so I/O does not
+/// hold the store mutex. The DB delete happens last; if disk removal fails
+/// the row stays so the next scan can recover.
+pub fn delete_extension(
+    store: &Mutex<Store>,
+    adapters: &[Box<dyn AgentAdapter>],
+    id: &str,
+) -> Result<(), HkError> {
+    // Phase 1: read metadata under the lock, then drop it before any I/O.
+    let (ext, projects) = {
+        let store = store.lock();
+        let ext = store
+            .get_extension(id)?
+            .ok_or_else(|| HkError::NotFound("Extension not found".into()))?;
+        let projects = store.list_project_tuples();
+        (ext, projects)
+    };
+
+    // Phase 2: filesystem / config-file mutation. No DB access here.
+    match ext.kind {
+        ExtensionKind::Skill => {
+            if let Some(loc) = scanner::find_skill_by_id(adapters, id, &ext.agents, &projects) {
+                if loc.entry_path.is_dir() {
+                    std::fs::remove_dir_all(&loc.entry_path)?;
+                } else {
+                    std::fs::remove_file(&loc.entry_path)?;
+                }
+            }
+        }
+        ExtensionKind::Mcp => {
+            for adapter in adapters.iter() {
+                if !ext.agents.contains(&adapter.name().to_string()) {
+                    continue;
+                }
+                let Some(config_path) = adapter.mcp_config_path_for(&ext.scope) else {
+                    continue;
+                };
+                for server in adapter.read_mcp_servers_from(&config_path) {
+                    let candidate = scanner::stable_id_with_scope_for(
+                        &server.name,
+                        "mcp",
+                        adapter.name(),
+                        &ext.scope,
+                    );
+                    if candidate == id {
+                        deployer::remove_mcp_server(
+                            &config_path,
+                            &server.name,
+                            adapter.mcp_format(),
+                        )?;
+                    }
+                }
+            }
+        }
+        ExtensionKind::Hook => {
+            for adapter in adapters.iter() {
+                if !ext.agents.contains(&adapter.name().to_string()) {
+                    continue;
+                }
+                let Some(config_path) = adapter.hook_config_path_for(&ext.scope) else {
+                    continue;
+                };
+                for hook in adapter.read_hooks_from(&config_path) {
+                    let hook_name = format!(
+                        "{}:{}:{}",
+                        hook.event,
+                        hook.matcher.as_deref().unwrap_or("*"),
+                        hook.command
+                    );
+                    let candidate = scanner::stable_id_with_scope_for(
+                        &hook_name,
+                        "hook",
+                        adapter.name(),
+                        &ext.scope,
+                    );
+                    if candidate == id {
+                        deployer::remove_hook(
+                            &config_path,
+                            &hook.event,
+                            hook.matcher.as_deref(),
+                            &hook.command,
+                            adapter.hook_format(),
+                        )?;
+                    }
+                }
+            }
+        }
+        ExtensionKind::Cli => {
+            // Child skills/MCPs are deleted separately by their own IDs.
+            // This branch only runs for full CLI uninstall (parent record cleanup).
+        }
+        ExtensionKind::Plugin => {
+            for adapter in adapters.iter() {
+                if !ext.agents.contains(&adapter.name().to_string()) {
+                    continue;
+                }
+                for plugin in adapter.read_plugins() {
+                    if scanner::stable_id_for(
+                        &format!("{}:{}", plugin.name, plugin.source),
+                        "plugin",
+                        adapter.name(),
+                    ) != id
+                    {
+                        continue;
+                    }
+                    let plugin_key = if plugin.source.is_empty() {
+                        plugin.name.clone()
+                    } else {
+                        format!("{}@{}", plugin.name, plugin.source)
+                    };
+                    if adapter.name() == "claude" {
+                        let config_path = adapter.plugin_config_path();
+                        deployer::remove_plugin_entry(&config_path, &plugin_key)?;
+                    } else if adapter.name() == "codex" {
+                        // Remove folder + config.toml entry
+                        if let Some(ref path) = plugin.path {
+                            let target = if let Some(parent) = path.parent() {
+                                if parent
+                                    .file_name()
+                                    .map(|n| n != "cache" && n != "plugins")
+                                    .unwrap_or(false)
+                                {
+                                    parent
+                                } else {
+                                    path.as_path()
+                                }
+                            } else {
+                                path.as_path()
+                            };
+                            if target.is_dir() {
+                                std::fs::remove_dir_all(target)?;
+                            } else if target.is_file() {
+                                std::fs::remove_file(target)?;
+                            }
+                        }
+                        deployer::remove_codex_plugin_entry(
+                            &adapter.mcp_config_path(),
+                            &plugin_key,
+                        )?;
+                    } else if adapter.name() == "gemini" {
+                        if let Some(ref path) = plugin.path
+                            && path.is_dir()
+                        {
+                            std::fs::remove_dir_all(path)?;
+                        }
+                        deployer::remove_gemini_extension_entry(
+                            &adapter.base_dir().join("extensions"),
+                            &plugin.name,
+                        )?;
+                    } else if adapter.name() == "copilot" {
+                        if let Some(ref path) = plugin.path
+                            && path.is_dir()
+                        {
+                            std::fs::remove_dir_all(path)?;
+                        }
+                        if let (Some(uri), Some(vscode_dir)) =
+                            (&plugin.uri, adapter.vscode_user_dir())
+                        {
+                            // Best-effort: VS Code may hold a lock on state.vscdb
+                            if let Err(e) =
+                                deployer::remove_vscode_plugin_entry(&vscode_dir, uri)
+                            {
+                                eprintln!(
+                                    "Warning: failed to clean up VS Code plugin entry: {e}"
+                                );
+                            }
+                        }
+                    } else if let Some(ref path) = plugin.path
+                        && path.is_dir()
+                    {
+                        // Cursor, etc. — just remove folder
+                        std::fs::remove_dir_all(path)?;
+                    }
+                }
+            }
+        }
+    }
+
+    // Phase 3: DB delete, only after disk side succeeded.
+    store.lock().delete_extension(id)
+}
+
+/// Read the rich on-disk content for an extension (skill text, MCP server
+/// config summary, hook detail, plugin README, …). Pure read-only — locks
+/// the store only to fetch metadata, then releases before any I/O.
+pub fn get_extension_content(
+    store: &Mutex<Store>,
+    adapters: &[Box<dyn AgentAdapter>],
+    id: &str,
+) -> Result<ExtensionContent, HkError> {
+    let (ext, projects) = {
+        let store = store.lock();
+        let ext = store
+            .get_extension(id)?
+            .ok_or_else(|| HkError::NotFound("Extension not found".into()))?;
+        let projects = store.list_project_tuples();
+        (ext, projects)
+    };
+
+    match ext.kind {
+        ExtensionKind::Skill => {
+            if let Some(loc) = scanner::find_skill_by_id(adapters, id, &ext.agents, &projects) {
+                let dir = if loc.entry_path.is_dir() {
+                    loc.entry_path.to_string_lossy().to_string()
+                } else {
+                    loc.skill_file
+                        .parent()
+                        .map(|p| p.to_string_lossy().to_string())
+                        .unwrap_or_default()
+                };
+                // Detect symlink: check entry itself, then parent skill_dir
+                let dir_symlink_target = if loc
+                    .skill_dir
+                    .symlink_metadata()
+                    .map(|m| m.is_symlink())
+                    .unwrap_or(false)
+                {
+                    std::fs::read_link(&loc.skill_dir).ok()
+                } else {
+                    None
+                };
+                let symlink_target = if loc
+                    .entry_path
+                    .symlink_metadata()
+                    .map(|m| m.is_symlink())
+                    .unwrap_or(false)
+                {
+                    std::fs::read_link(&loc.entry_path)
+                        .ok()
+                        .map(|t| t.to_string_lossy().to_string())
+                } else if let Some(ref resolved_dir) = dir_symlink_target {
+                    let entry_name = loc.entry_path.file_name().unwrap_or_default();
+                    Some(resolved_dir.join(entry_name).to_string_lossy().to_string())
+                } else {
+                    None
+                };
+                let content = std::fs::read_to_string(&loc.skill_file)?;
+                Ok(ExtensionContent {
+                    content,
+                    path: Some(dir),
+                    symlink_target,
+                })
+            } else {
+                Err(HkError::NotFound("Skill file not found".into()))
+            }
+        }
+        ExtensionKind::Mcp => {
+            // The trait helper resolves the right file per scope; the
+            // scanner's `source_path` is the canonical config path for project
+            // entries — we prefer it when set.
+            let mut fallback_config_path = ext.source_path.clone();
+            for adapter in adapters {
+                if !ext.agents.contains(&adapter.name().to_string()) {
+                    continue;
+                }
+                let Some(config_path) = adapter.mcp_config_path_for(&ext.scope) else {
+                    continue;
+                };
+                if fallback_config_path.is_none() {
+                    fallback_config_path = Some(config_path.to_string_lossy().to_string());
+                }
+                for server in adapter.read_mcp_servers_from(&config_path) {
+                    let candidate = scanner::stable_id_with_scope_for(
+                        &server.name,
+                        "mcp",
+                        adapter.name(),
+                        &ext.scope,
+                    );
+                    if candidate == id {
+                        let mut lines = vec![format!("Command: {}", server.command)];
+                        if !server.args.is_empty() {
+                            lines.push(format!("Args: {}", server.args.join(" ")));
+                        }
+                        if !server.env.is_empty() {
+                            lines.push("Environment:".into());
+                            for k in server.env.keys() {
+                                lines.push(format!("  {} = ****", k));
+                            }
+                        }
+                        return Ok(ExtensionContent {
+                            content: lines.join("\n"),
+                            path: Some(config_path.to_string_lossy().to_string()),
+                            symlink_target: None,
+                        });
+                    }
+                }
+            }
+            // Disabled MCP: still surface the config path where it lived.
+            Ok(ExtensionContent {
+                content: ext.description,
+                path: fallback_config_path,
+                symlink_target: None,
+            })
+        }
+        ExtensionKind::Hook => {
+            let mut fallback_config_path = ext.source_path.clone();
+            for adapter in adapters {
+                if !ext.agents.contains(&adapter.name().to_string()) {
+                    continue;
+                }
+                let Some(config_path) = adapter.hook_config_path_for(&ext.scope) else {
+                    continue;
+                };
+                if fallback_config_path.is_none() {
+                    fallback_config_path = Some(config_path.to_string_lossy().to_string());
+                }
+                for hook in adapter.read_hooks_from(&config_path) {
+                    let hook_name = format!(
+                        "{}:{}:{}",
+                        hook.event,
+                        hook.matcher.as_deref().unwrap_or("*"),
+                        hook.command
+                    );
+                    let candidate = scanner::stable_id_with_scope_for(
+                        &hook_name,
+                        "hook",
+                        adapter.name(),
+                        &ext.scope,
+                    );
+                    if candidate == id {
+                        let mut lines = vec![format!("Event: {}", hook.event)];
+                        if let Some(m) = &hook.matcher {
+                            lines.push(format!("Matcher: {}", m));
+                        }
+                        lines.push(format!("Command: {}", hook.command));
+                        return Ok(ExtensionContent {
+                            content: lines.join("\n"),
+                            path: Some(config_path.to_string_lossy().to_string()),
+                            symlink_target: None,
+                        });
+                    }
+                }
+            }
+            Ok(ExtensionContent {
+                content: ext.description,
+                path: fallback_config_path,
+                symlink_target: None,
+            })
+        }
+        ExtensionKind::Plugin => {
+            for adapter in adapters {
+                if !ext.agents.contains(&adapter.name().to_string()) {
+                    continue;
+                }
+                for plugin in adapter.read_plugins() {
+                    if scanner::stable_id_for(
+                        &format!("{}:{}", plugin.name, plugin.source),
+                        "plugin",
+                        adapter.name(),
+                    ) == id
+                    {
+                        let path_str = plugin
+                            .path
+                            .as_ref()
+                            .map(|p| p.to_string_lossy().to_string());
+                        // Try README.md from plugin dir first, then walk up to
+                        // find the repo root (for git-cloned plugins where
+                        // README sits one or more levels above the manifest).
+                        let content = plugin
+                            .path
+                            .as_ref()
+                            .and_then(|p| {
+                                for candidate in [p.join("README.md"), p.join("readme.md")] {
+                                    if let Ok(text) = std::fs::read_to_string(&candidate) {
+                                        return Some(text);
+                                    }
+                                }
+                                let mut dir = p.clone();
+                                while dir.pop() {
+                                    if dir.join(".git").exists() {
+                                        for name in ["README.md", "readme.md"] {
+                                            if let Ok(text) =
+                                                std::fs::read_to_string(dir.join(name))
+                                            {
+                                                return Some(text);
+                                            }
+                                        }
+                                        break;
+                                    }
+                                }
+                                None
+                            })
+                            .unwrap_or(ext.description);
+                        return Ok(ExtensionContent {
+                            content,
+                            path: path_str,
+                            symlink_target: None,
+                        });
+                    }
+                }
+            }
+            Ok(ExtensionContent {
+                content: ext.description,
+                path: None,
+                symlink_target: None,
+            })
+        }
+        ExtensionKind::Cli => Ok(ExtensionContent {
+            content: ext.description,
+            path: None,
+            symlink_target: None,
+        }),
+    }
+}
+
+/// Cross-agent deploy: copy a Skill / MCP / Hook / CLI from its source agent
+/// into `target_agent`. Returns a human-readable identifier of what was
+/// deployed (skill name, MCP server name, or `event:command` for hooks) so
+/// the UI can show the result. The wrapper is responsible for any post-deploy
+/// rescan/sync (web does this; desktop does not, matching prior behavior).
+pub fn install_to_agent(
+    store: &Mutex<Store>,
+    adapters: &[Box<dyn AgentAdapter>],
+    extension_id: &str,
+    target_agent: &str,
+) -> Result<String, HkError> {
+    let (ext, projects) = {
+        let store = store.lock();
+        let ext = store
+            .get_extension(extension_id)?
+            .ok_or_else(|| HkError::NotFound("Extension not found".into()))?;
+        let projects = store.list_project_tuples();
+        (ext, projects)
+    };
+
+    let target_adapter = adapters
+        .iter()
+        .find(|a| a.name() == target_agent)
+        .ok_or_else(|| HkError::NotFound(format!("Agent '{}' not found", target_agent)))?;
+
+    match ext.kind {
+        ExtensionKind::Skill => {
+            let source_path =
+                scanner::find_skill_by_id(adapters, extension_id, &ext.agents, &projects)
+                    .map(|loc| loc.entry_path)
+                    .ok_or_else(|| {
+                        HkError::Internal("Could not find source skill files".into())
+                    })?;
+            let target_dir = target_adapter
+                .skill_dirs()
+                .into_iter()
+                .next()
+                .ok_or_else(|| {
+                    HkError::Internal(format!(
+                        "No skill directory for agent '{}'",
+                        target_agent
+                    ))
+                })?;
+            let deployed_name = deployer::deploy_skill(&source_path, &target_dir)?;
+            Ok(deployed_name)
+        }
+        ExtensionKind::Mcp => {
+            let mut source_entry = None;
+            for adapter in adapters.iter() {
+                if !ext.agents.contains(&adapter.name().to_string()) {
+                    continue;
+                }
+                let Some(source_path) = adapter.mcp_config_path_for(&ext.scope) else {
+                    continue;
+                };
+                for server in adapter.read_mcp_servers_from(&source_path) {
+                    let candidate = scanner::stable_id_with_scope_for(
+                        &server.name,
+                        "mcp",
+                        adapter.name(),
+                        &ext.scope,
+                    );
+                    if candidate == extension_id {
+                        source_entry = Some(server);
+                        break;
+                    }
+                }
+                if source_entry.is_some() {
+                    break;
+                }
+            }
+            let mut entry = source_entry.ok_or_else(|| {
+                HkError::Internal("Could not find source MCP server config".into())
+            })?;
+            if target_adapter.needs_path_injection() {
+                deployer::ensure_path_injection(&mut entry);
+            }
+            let config_path = target_adapter.mcp_config_path();
+            deployer::deploy_mcp_server(&config_path, &entry, target_adapter.mcp_format())?;
+            Ok(entry.name)
+        }
+        ExtensionKind::Hook => {
+            let mut source_entry = None;
+            for adapter in adapters.iter() {
+                if !ext.agents.contains(&adapter.name().to_string()) {
+                    continue;
+                }
+                let Some(source_path) = adapter.hook_config_path_for(&ext.scope) else {
+                    continue;
+                };
+                for hook in adapter.read_hooks_from(&source_path) {
+                    let hook_name = format!(
+                        "{}:{}:{}",
+                        hook.event,
+                        hook.matcher.as_deref().unwrap_or("*"),
+                        hook.command
+                    );
+                    let candidate = scanner::stable_id_with_scope_for(
+                        &hook_name,
+                        "hook",
+                        adapter.name(),
+                        &ext.scope,
+                    );
+                    if candidate == extension_id {
+                        source_entry = Some(hook);
+                        break;
+                    }
+                }
+                if source_entry.is_some() {
+                    break;
+                }
+            }
+            let mut entry = source_entry
+                .ok_or_else(|| HkError::Internal("Could not find source hook config".into()))?;
+
+            // Translate event name to the target agent's convention. Agents
+            // disagree on hook event names (Claude `PreToolUse` vs Codex
+            // `pre_tool_use`, etc.) so a missing translation is a hard error.
+            let translated_event = target_adapter
+                .translate_hook_event(&entry.event)
+                .ok_or_else(|| {
+                    HkError::Internal(format!(
+                        "Hook event '{}' is not supported by {}",
+                        entry.event, target_agent
+                    ))
+                })?;
+            entry.event = translated_event;
+
+            let config_path = target_adapter.hook_config_path();
+            deployer::deploy_hook(&config_path, &entry, target_adapter.hook_format())?;
+
+            // Codex requires hooks feature enabled in config.toml
+            if target_adapter.name() == "codex"
+                && let Err(e) = deployer::ensure_codex_hooks_enabled(&target_adapter.base_dir())
+            {
+                eprintln!("[hk] warning: {e}");
+            }
+
+            Ok(format!("{}:{}", entry.event, entry.command))
+        }
+        ExtensionKind::Cli => {
+            // Deploy the CLI's associated skill to the target agent.
+            let binary_name = ext
+                .cli_meta
+                .as_ref()
+                .map(|m| m.binary_name.clone())
+                .unwrap_or_else(|| ext.name.to_lowercase());
+            let locations = scanner::skill_locations(&binary_name, adapters, &projects);
+            let source_path = locations
+                .into_iter()
+                .next()
+                .map(|(_, path)| path)
+                .ok_or_else(|| {
+                    HkError::Internal("Could not find source skill files for CLI".into())
+                })?;
+            let target_dir = target_adapter
+                .skill_dirs()
+                .into_iter()
+                .next()
+                .ok_or_else(|| {
+                    HkError::Internal(format!(
+                        "No skill directory for agent '{}'",
+                        target_agent
+                    ))
+                })?;
+            let deployed_name = deployer::deploy_skill(&source_path, &target_dir)?;
+            Ok(deployed_name)
+        }
+        other => Err(HkError::Internal(format!(
+            "Cross-agent deploy not supported for '{}' extensions",
+            other.as_str()
+        ))),
+    }
 }
 
 #[cfg(test)]
