+++ crates/hk-desktop/src/commands/helpers.rs
@@ -1,129 +1,4 @@
-use hk_core::{
-    HkError, adapter,
-    auditor::{self, Auditor},
-    models::*,
-    scanner,
-};
-use std::path::PathBuf;
-
-/// Information about a skill's physical location on disk.
-pub(super) struct SkillLocation {
-    /// The filesystem entry (directory or .md file).
-    pub(super) entry_path: PathBuf,
-    /// The SKILL.md file (same as entry_path for standalone .md files).
-    pub(super) skill_file: PathBuf,
-    /// The parent skill directory that contained this entry.
-    pub(super) skill_dir: PathBuf,
-}
-
-/// Scan all `adapters` for the skill whose stable ID matches `ext_id`,
-/// restricting to adapters whose names appear in `agent_filter`.
-///
-/// Returns `Some(SkillLocation)` on the first match, or `None`.
-pub(super) fn find_skill_by_id(
-    adapters: &[Box<dyn adapter::AgentAdapter>],
-    ext_id: &str,
-    agent_filter: &[String],
-) -> Option<SkillLocation> {
-    for a in adapters {
-        if !agent_filter.contains(&a.name().to_string()) {
-            continue;
-        }
-        for skill_dir in a.skill_dirs() {
-            let Ok(entries) = std::fs::read_dir(&skill_dir) else {
-                continue;
-            };
-            for entry in entries.flatten() {
-                let path = entry.path();
-                let skill_file = if path.is_dir() {
-                    let md = path.join("SKILL.md");
-                    if md.exists() {
-                        md
-                    } else {
-                        path.join("SKILL.md.disabled")
-                    }
-                } else if path
-                    .extension()
-                    .is_some_and(|e| e == "md" || e == "disabled")
-                {
-                    path.clone()
-                } else {
-                    continue;
-                };
-                if !skill_file.exists() {
-                    continue;
-                }
-                let name = scanner::parse_skill_name(&skill_file).unwrap_or_else(|| {
-                    path.file_stem()
-                        .unwrap_or_default()
-                        .to_string_lossy()
-                        .to_string()
-                });
-                if scanner::stable_id_for(&name, "skill", a.name()) == ext_id {
-                    return Some(SkillLocation {
-                        entry_path: path,
-                        skill_file,
-                        skill_dir: skill_dir.clone(),
-                    });
-                }
-            }
-        }
-    }
-    None
-}
-
-/// Audit a single extension by name (best-effort).
-/// Takes owned extensions list so the caller can release the Mutex before calling this.
-/// Returns audit results to be stored by the caller.
-/// NOTE: Superseded by `hk_core::service::post_install_sync` which includes audit internally.
-#[allow(dead_code)]
-pub(super) fn audit_extension_by_name(
-    name: &str,
-    extensions: &[Extension],
-    adapters: &[Box<dyn adapter::AgentAdapter>],
-) -> Vec<AuditResult> {
-    let auditor = Auditor::new();
-    let mut results = Vec::new();
-    for ext in extensions {
-        if ext.name != name {
-            continue;
-        }
-        let input = match ext.kind {
-            ExtensionKind::Skill => {
-                let (content, file_path) =
-                    if let Some(loc) = find_skill_by_id(adapters, &ext.id, &ext.agents) {
-                        (
-                            std::fs::read_to_string(&loc.skill_file).unwrap_or_default(),
-                            loc.skill_file.to_string_lossy().to_string(),
-                        )
-                    } else {
-                        (String::new(), ext.name.clone())
-                    };
-                auditor::AuditInput {
-                    extension_id: ext.id.clone(),
-                    kind: ext.kind,
-                    name: ext.name.clone(),
-                    content,
-                    source: ext.source.clone(),
-                    file_path,
-                    mcp_command: None,
-                    mcp_args: vec![],
-                    mcp_env: Default::default(),
-                    installed_at: ext.installed_at,
-                    updated_at: ext.updated_at,
-                    permissions: ext.permissions.clone(),
-                    cli_parent_id: ext.cli_parent_id.clone(),
-                    cli_meta: ext.cli_meta.clone(),
-                    child_permissions: vec![],
-                    pack: ext.pack.clone(),
-                }
-            }
-            _ => continue,
-        };
-        results.push(auditor.audit(&input));
-    }
-    results
-}
+use hk_core::HkError;
 
 /// Validate that a path is within a known agent directory, registered project, or the app data dir.
 pub(super) fn is_path_within_allowed_dirs(
