+++ crates/hk-desktop/src/commands/install.rs
@@ -1,4 +1,3 @@
-use super::helpers::find_skill_by_id;
 use super::{AppState, PendingClone};
 use hk_core::{HkError, deployer, manager, marketplace, models::*, scanner, service};
 use tauri::State;
@@ -533,157 +532,10 @@ pub async fn install_to_agent(
     extension_id: String,
     target_agent: String,
 ) -> Result<String, HkError> {
-    let store_clone = state.store.clone();
+    let store = state.store.clone();
     let adapters = state.adapters.clone();
     tauri::async_runtime::spawn_blocking(move || {
-        let ext = {
-            let store = store_clone.lock();
-            store
-                .get_extension(&extension_id)?
-                .ok_or_else(|| HkError::NotFound("Extension not found".into()))?
-        };
-
-        let target_adapter = adapters
-            .iter()
-            .find(|a| a.name() == target_agent)
-            .ok_or_else(|| HkError::NotFound(format!("Agent '{}' not found", target_agent)))?;
-
-        match ext.kind {
-            ExtensionKind::Skill => {
-                // Find source skill path
-                let source_path = find_skill_by_id(&adapters, &extension_id, &ext.agents)
-                    .map(|loc| loc.entry_path)
-                    .ok_or_else(|| {
-                        HkError::Internal("Could not find source skill files".into())
-                    })?;
-                let target_dir = target_adapter
-                    .skill_dirs()
-                    .into_iter()
-                    .next()
-                    .ok_or_else(|| {
-                        HkError::Internal(format!(
-                            "No skill directory for agent '{}'",
-                            target_agent
-                        ))
-                    })?;
-                let deployed_name = deployer::deploy_skill(&source_path, &target_dir)?;
-                Ok(deployed_name)
-            }
-            ExtensionKind::Mcp => {
-                // Find the source MCP server entry
-                let mut source_entry = None;
-                for adapter in adapters.iter() {
-                    if !ext.agents.contains(&adapter.name().to_string()) {
-                        continue;
-                    }
-                    for server in adapter.read_mcp_servers() {
-                        if scanner::stable_id_for(&server.name, "mcp", adapter.name())
-                            == extension_id
-                        {
-                            source_entry = Some(server);
-                            break;
-                        }
-                    }
-                    if source_entry.is_some() {
-                        break;
-                    }
-                }
-                let mut entry = source_entry.ok_or_else(|| {
-                    HkError::Internal("Could not find source MCP server config".into())
-                })?;
-                if target_adapter.needs_path_injection() {
-                    deployer::ensure_path_injection(&mut entry);
-                }
-                let config_path = target_adapter.mcp_config_path();
-                deployer::deploy_mcp_server(&config_path, &entry, target_adapter.mcp_format())?;
-                Ok(entry.name)
-            }
-            ExtensionKind::Hook => {
-                // Find the source hook entry
-                let mut source_entry = None;
-                for adapter in adapters.iter() {
-                    if !ext.agents.contains(&adapter.name().to_string()) {
-                        continue;
-                    }
-                    for hook in adapter.read_hooks() {
-                        let hook_name = format!(
-                            "{}:{}:{}",
-                            hook.event,
-                            hook.matcher.as_deref().unwrap_or("*"),
-                            hook.command
-                        );
-                        if scanner::stable_id_for(&hook_name, "hook", adapter.name())
-                            == extension_id
-                        {
-                            source_entry = Some(hook);
-                            break;
-                        }
-                    }
-                    if source_entry.is_some() {
-                        break;
-                    }
-                }
-                let mut entry = source_entry.ok_or_else(|| {
-                    HkError::Internal("Could not find source hook config".into())
-                })?;
-
-                // Translate event name to target agent's convention
-                let translated_event = target_adapter
-                    .translate_hook_event(&entry.event)
-                    .ok_or_else(|| {
-                        HkError::Internal(format!(
-                            "Hook event '{}' is not supported by {}",
-                            entry.event, target_agent
-                        ))
-                    })?;
-                entry.event = translated_event;
-
-                let config_path = target_adapter.hook_config_path();
-                deployer::deploy_hook(&config_path, &entry, target_adapter.hook_format())?;
-
-                // Codex requires hooks feature enabled in config.toml
-                if target_adapter.name() == "codex"
-                    && let Err(e) =
-                        deployer::ensure_codex_hooks_enabled(&target_adapter.base_dir())
-                {
-                    eprintln!("[hk] warning: {e}");
-                }
-
-                Ok(format!("{}:{}", entry.event, entry.command))
-            }
-            ExtensionKind::Cli => {
-                // Deploy the CLI's associated skill to the target agent
-                let binary_name = ext
-                    .cli_meta
-                    .as_ref()
-                    .map(|m| m.binary_name.clone())
-                    .unwrap_or_else(|| ext.name.to_lowercase());
-                let locations = scanner::skill_locations(&binary_name, &adapters);
-                let source_path = locations
-                    .into_iter()
-                    .next()
-                    .map(|(_, path)| path)
-                    .ok_or_else(|| {
-                        HkError::Internal("Could not find source skill files for CLI".into())
-                    })?;
-                let target_dir = target_adapter
-                    .skill_dirs()
-                    .into_iter()
-                    .next()
-                    .ok_or_else(|| {
-                        HkError::Internal(format!(
-                            "No skill directory for agent '{}'",
-                            target_agent
-                        ))
-                    })?;
-                let deployed_name = deployer::deploy_skill(&source_path, &target_dir)?;
-                Ok(deployed_name)
-            }
-            other => Err(HkError::Internal(format!(
-                "Cross-agent deploy not supported for '{}' extensions",
-                other.as_str()
-            ))),
-        }
+        service::install_to_agent(&store, &adapters, &extension_id, &target_agent)
     })
     .await
     .map_err(|e| HkError::Internal(e.to_string()))?
